feat: extend journal entry schema

+++ services/zoe-core/main.py
@@ -294,7 +294,7 @@ class JournalEntry(BaseModel):
     content: str
     mood: Optional[str] = None
     mood_score: Optional[float] = None
-    mood_confidence: Optional[int] = None
+    mood_confidence: Optional[float] = None
     photo: Optional[str] = None
     health_info: Optional[dict] = None
     user_id: str = "default"
@@ -305,7 +305,7 @@ class JournalEntryResponse(BaseModel):
     content: str
     mood: Optional[str]
     mood_score: Optional[float]
-    mood_confidence: Optional[int]
+    mood_confidence: Optional[float]
     photo: Optional[str]
     health_info: Optional[dict]
     created_at: str
@@ -317,7 +317,7 @@ class JournalEntryResponse(BaseModel):
     content: str
     mood: Optional[str]
     mood_score: Optional[float]
-    mood_confidence: Optional[int]
+    mood_confidence: Optional[float]
     photo: Optional[str]
     health_info: Optional[dict]
